requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
joblib==1.1.0
pandas==1.2.1
numpy==1.19.5
requests==2.22.0
sklearn